Update README.md
The argument n is in x0. I used offsets by factors of eight to store the temporary registers. They are then loaded again later in the Fibonacci function. The code uses recursion to continuously add to n until it represents the nth term in the Fibonacci sequence. During the recursion, the temporary value is stored in x10 and new arguments are given to x1 and x2. Once n (x0) gets down to 1, the will end giving the nth term of the Fibonacci sequence. At the end, there is a print function to display the value.